cs-churn-predictor/requirements.txt
# CS Churn Predictor — dependencies

# Data & ML
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
joblib>=1.3

# API
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.0

# Visualization (notebooks)
matplotlib>=3.7
seaborn>=0.13
jupyter>=1.0
ipykernel>=6.0

# Policy engine (rule loading from YAML)
PyYAML>=6.0

# Optional: faster gradient boosting
# xgboost>=2.0
# lightgbm>=4.0
